Replace dockerize by confd for templating
dockerize tries to chmod the file after generation, which it couldn't
as now we have restricted access on /etc/odoo.cfg

Note: with confd, I had to use a workaround which is to generate the
file in /tmp then copy it's content in /etc/odoo.cfg (the user doesn't
have write perms in /etc, only on /etc/odoo.cfg and confd creates a
temporary file in the same directory to avoid moving files between
partitions). I could have done the same trick with dockerize then, but
the switch was already done.